Add hyperfoil
