+++ quarkus-modules/quarkus-vs-springboot/hyperfoil/volume/cities.csv
@@ -1,0 +1,136 @@
+Holtsville
+Adjuntas
+Aguada
+Aguadilla
+Maricao
+Anasco
+Angeles
+Arecibo
+Bajadero
+Barceloneta
+Boqueron
+Cabo Rojo
+Penuelas
+Camuy
+Castaner
+Rosario
+Sabana Grande
+Ciales
+Utuado
+Dorado
+Ensenada
+Florida
+Garrochales
+Guanica
+Guayanilla
+Hatillo
+Hormigueros
+Isabela
+Jayuya
+Lajas
+Lares
+Las Marias
+Manati
+Moca
+Rincon
+Quebradillas
+Mayaguez
+San German
+San Sebastian
+Morovis
+Sabana Hoyos
+San Antonio
+Vega Alta
+Vega Baja
+Yauco
+Aguas Buenas
+Aguirre
+Aibonito
+Maunabo
+Arroyo
+Mercedita
+Ponce
+Naguabo
+Naranjito
+Orocovis
+Palmer
+Patillas
+Caguas
+Canovanas
+Ceiba
+Cayey
+Fajardo
+Cidra
+Puerto Real
+Punta Santiago
+Roosevelt Roads
+Rio Blanco
+Rio Grande
+Salinas
+San Lorenzo
+Santa Isabel
+Vieques
+Villalba
+Yabucoa
+Coamo
+Las Piedras
+Loiza
+Luquillo
+Culebra
+Juncos
+Gurabo
+Coto Laurel
+Comerio
+Corozal
+Guayama
+La Plata
+Humacao
+Barranquitas
+Juana Diaz
+St Thomas
+Christiansted
+St John
+Frederiksted
+Kingshill
+San Juan
+Fort Buchanan
+Toa Baja
+Sabana Seca
+Toa Alta
+Bayamon
+Catano
+Guaynabo
+Trujillo Alto
+Saint Just
+Carolina
+Agawam
+Amherst
+Barre
+Belchertown
+Blandford
+Bondsville
+Brimfield
+Chester
+Chesterfield
+Chicopee
+Cummington
+Easthampton
+East Longmeadow
+East Otis
+Feeding Hills
+Gilbertville
+Goshen
+Granby
+Granville
+Hadley
+Hampden
+Hardwick
+Hatfield
+Haydenville
+Holyoke
+Huntington
+Leeds
+Leverett
+Ludlow
+Monson
+North Amherst
