--- quarkus-modules/quarkus-vs-springboot/run_test_jmeter.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-$Jmeter_home/bin/jmeter -n -t load_test.jmx -l log.csv -e -o ./report
