+++ quarkus-modules/quarkus-vs-springboot/spring-project/src/main/docker/spring.yml
@@ -1,27 +1,25 @@
 version: '3.1'
 
 services:
-  #db:
-  #  image: mysql:5.7.38
-  #  ports:
-  #    - '3306:3306'
-  #  environment:
-  #    MYSQL_ROOT_PASSWORD: root
-  #    MYSQL_DATABASE: baeldung
-  #  command: [ 'mysqld', '--character-set-server=utf8mb4', '--collation-server=utf8mb4_unicode_ci' ]
-  #  healthcheck:
-  #    test: mysqladmin ping -h 127.0.0.1 -u $$MYSQL_USER --password=$$MYSQL_PASSWORD
+  db:
+    image: mysql:5.7.38
+    ports:
+      - '3306:3306'
+    environment:
+      MYSQL_ROOT_PASSWORD: root
+      MYSQL_DATABASE: baeldung
+    command: [ 'mysqld', '--character-set-server=utf8mb4', '--collation-server=utf8mb4_unicode_ci' ]
+    healthcheck:
+      test: mysqladmin ping -h 127.0.0.1 -u $$MYSQL_USER --password=$$MYSQL_PASSWORD
   app:
     image: docker.io/library/spring-project:0.1-SNAPSHOT
     network_mode: "host"
     environment:
-      DB_URL: r2dbc:mysql://${EXTERNAL_IP}:3306/baeldung?useSSL=true&requireSSL=true
+      DB_URL: r2dbc:mysql://localhost:3306/baeldung?useSSL=true&requireSSL=true
       HOST_HOSTNAME: ${EXTERNAL_IP}
-        # links:
-    #  - "db"
-      #depends_on:
-      #db:
-    #  condition: service_healthy
+    depends_on:
+      db:
+        condition: service_healthy
     deploy:
       resources:
         limits:
