+++ quarkus-modules/quarkus-vs-springboot/wrk/generator.lua
@@ -12,7 +12,6 @@ function ParseCSVLine(line,sep)
         local c = string.sub(line,pos,pos)
         if (c == "") then break end
         if (c == '"') then
-            -- quoted value (ignore separator within)
             local txt = ""
             repeat
                 local startp,endp = string.find(line,'^%b""',pos)
@@ -20,22 +19,16 @@ function ParseCSVLine(line,sep)
                 pos = endp + 1
                 c = string.sub(line,pos,pos)
                 if (c == '"') then txt = txt..'"' end
-                -- check first char AFTER quoted string, if it is another
-                -- quoted string without separator, then append it
-                -- this is the way to "escape" the quote char in a quote. example:
-                --   value1,"blub""blip""boing",value3  will result in blub"blip"boing  for the middle
             until (c ~= '"')
             table.insert(res,txt)
             assert(c == sep or c == "")
             pos = pos + 1
         else
-            -- no quotes used, just look for the first separator
             local startp,endp = string.find(line,sep,pos)
             if (startp) then
                 table.insert(res,string.sub(line,pos,startp-1))
                 pos = endp + 1
             else
-                -- no separator found -> use rest of string and terminate
                 table.insert(res,string.sub(line,pos))
                 break
             end
@@ -45,10 +38,8 @@ function ParseCSVLine(line,sep)
 end
 
 loadFile = function()
-    -- 1. Get file path
     local filename = "zip_code_database.csv"
 
-    -- 3. Parse the data:
     local data = {}
     local count = 0
     local sep = ","
