+++ quarkus-modules/quarkus-vs-springboot/wrk/json.lua
@@ -1,9 +1,5 @@
--- https://gist.github.com/tylerneylon/59f4bcf316be525b30ab
-
 local json = {}
 
-
--- Internal functions.
 local function kind_of(obj)
   if type(obj) ~= 'table' then return type(obj) end
   local i = 1
@@ -22,10 +18,6 @@ local function escape_str(s)
   return s
 end
 
--- Returns pos, did_find; there are two cases:
--- 1. Delimiter found: pos = pos after leading space + delim; did_find = true.
--- 2. Delimiter not found: pos = pos after leading space;     did_find = false.
--- This throws an error if err_if_missing is true and the delim is not found.
 local function skip_delim(str, pos, delim, err_if_missing)
   pos = pos + #str:match('^%s*', pos)
   if str:sub(pos, pos) ~= delim then
@@ -37,8 +29,6 @@ local function skip_delim(str, pos, delim, err_if_missing)
   return pos + 1, true
 end
 
--- Expects the given pos to be the first character after the opening quote.
--- Returns val, pos; the returned pos is after the closing quote character.
 local function parse_str_val(str, pos, val)
   val = val or ''
   local early_end_error = 'End of input found while parsing string.'
@@ -46,14 +36,12 @@ local function parse_str_val(str, pos, val)
   local c = str:sub(pos, pos)
   if c == '"'  then return val, pos + 1 end
   if c ~= '\\' then return parse_str_val(str, pos + 1, val .. c) end
-  -- We must have a \ character.
   local esc_map = {b = '\b', f = '\f', n = '\n', r = '\r', t = '\t'}
   local nextc = str:sub(pos + 1, pos + 1)
   if not nextc then error(early_end_error) end
   return parse_str_val(str, pos + 2, val .. (esc_map[nextc] or nextc))
 end
 
--- Returns val, pos; the returned pos is after the number's final character.
 local function parse_num_val(str, pos)
   local num_str = str:match('^-?%d+%.?%d*[eE]?[+-]?%d*', pos)
   local val = tonumber(num_str)
@@ -61,11 +49,9 @@ local function parse_num_val(str, pos)
   return val, pos + #num_str
 end
 
--- Public values and functions.
-
 function json.stringify(obj, as_key)
-  local s = {}  -- We'll build the string as an array of strings to be concatenated.
-  local kind = kind_of(obj)  -- This is 'array' if it's an array or type(obj) otherwise.
+  local s = {}
+  local kind = kind_of(obj)
   if kind == 'array' then
     if as_key then error('Can\'t encode array as key.') end
     s[#s + 1] = '['
@@ -99,25 +85,25 @@ function json.stringify(obj, as_key)
   return table.concat(s)
 end
 
-json.null = {}  -- This is a one-off table to represent the null value.
+json.null = {}
 
 function json.parse(str, pos, end_delim)
   pos = pos or 1
   if pos > #str then error('Reached unexpected end of input.') end
-  local pos = pos + #str:match('^%s*', pos)  -- Skip whitespace.
+  local pos = pos + #str:match('^%s*', pos)
   local first = str:sub(pos, pos)
-  if first == '{' then  -- Parse an object.
+  if first == '{' then
     local obj, key, delim_found = {}, true, true
     pos = pos + 1
     while true do
       key, pos = json.parse(str, pos, '}')
       if key == nil then return obj, pos end
       if not delim_found then error('Comma missing between object items.') end
-      pos = skip_delim(str, pos, ':', true)  -- true -> error if missing.
+      pos = skip_delim(str, pos, ':', true)
       obj[key], pos = json.parse(str, pos)
       pos, delim_found = skip_delim(str, pos, ',')
     end
-  elseif first == '[' then  -- Parse an array.
+  elseif first == '[' then
     local arr, val, delim_found = {}, true, true
     pos = pos + 1
     while true do
@@ -127,13 +113,13 @@ function json.parse(str, pos, end_delim)
       arr[#arr + 1] = val
       pos, delim_found = skip_delim(str, pos, ',')
     end
-  elseif first == '"' then  -- Parse a string.
+  elseif first == '"' then
     return parse_str_val(str, pos + 1)
-  elseif first == '-' or first:match('%d') then  -- Parse a number.
+  elseif first == '-' or first:match('%d') then
     return parse_num_val(str, pos)
-  elseif first == end_delim then  -- End of an object or array.
+  elseif first == end_delim then
     return nil, pos + 1
-  else  -- Parse true, false, or null.
+  else
     local literals = {['true'] = true, ['false'] = false, ['null'] = json.null}
     for lit_str, lit_val in pairs(literals) do
       local lit_end = pos + #lit_str - 1
