+++ quarkus-modules/quarkus-vs-springboot/wrk/post_zipcode.lua
@@ -3,7 +3,6 @@ local json = require "json"
 
 math.randomseed(os.clock()*100000000000)
 
--- read csv lines
 function ParseCSVLine(line,sep)
 	local res = {}
 	local pos = 1
@@ -12,7 +11,6 @@ function ParseCSVLine(line,sep)
 		local c = string.sub(line,pos,pos)
 		if (c == "") then break end
 		if (c == '"') then
-			-- quoted value (ignore separator within)
 			local txt = ""
 			repeat
 				local startp,endp = string.find(line,'^%b""',pos)
@@ -20,22 +18,17 @@ function ParseCSVLine(line,sep)
 				pos = endp + 1
 				c = string.sub(line,pos,pos)
 				if (c == '"') then txt = txt..'"' end
-				-- check first char AFTER quoted string, if it is another
-				-- quoted string without separator, then append it
-				-- this is the way to "escape" the quote char in a quote. example:
-				--   value1,"blub""blip""boing",value3  will result in blub"blip"boing  for the middle
+
 			until (c ~= '"')
 			table.insert(res,txt)
 			assert(c == sep or c == "")
 			pos = pos + 1
 		else
-			-- no quotes used, just look for the first separator
 			local startp,endp = string.find(line,sep,pos)
 			if (startp) then
 				table.insert(res,string.sub(line,pos,startp-1))
 				pos = endp + 1
 			else
-				-- no separator found -> use rest of string and terminate
 				table.insert(res,string.sub(line,pos))
 				break
 			end
@@ -45,10 +38,8 @@ function ParseCSVLine(line,sep)
 end
 
 loadFile = function()
-    -- 1. Get file path
     local filename = "zip_code_database.csv"
 
-    -- 3. Parse the data:
     local data = {}
     local count = 0
     local sep = ","
@@ -72,8 +63,6 @@ request = function()
 	return wrk.format("POST", url_path, headers, json.stringify(val))
 end
 
--- example reporting script which demonstrates a custom
--- done() function that prints latency percentiles as CSV
 done = function(summary, latency, requests)
     io.write("--------------POST ZIPCODE----------------\n")
     for _, p in pairs({ 50, 75, 90, 99, 99.999 }) do
